Move memcache configuration to memcache backend
